% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{handle.missing}
\alias{handle.missing}
\title{Handle missing values

This function deals with the cleaning and/or imputation of the missing values in the data.}
\usage{
handle.missing(features, type)
}
\arguments{
\item{features}{A dataframe containing the dataset you want to clean or impute.}

\item{type}{A string which can take the value "remove" if you want to remove the nonsense 
0's in the data or "impute" if you want to impute them from the remaining observations.}
}
\value{
The features dataframe modified accordingly.
}

