% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lasso.R
\name{lasso.model}
\alias{lasso.model}
\title{Lasso model for feature selection and prediction}
\usage{
lasso.model(train.features, labels, type = "select", test.features = NULL,
  dfmax = NULL)
}
\arguments{
\item{train.features}{A dataframe containing the training dataset (output excluded).}

\item{labels}{A vector of multi-class labels for the training set.}

\item{type}{A string with values "select" for performing feature selection or "predict" for 
predictions using Lasso (by default it is set to "select").}

\item{test.features}{A dataframe containing the test dataset (NULL if type = "select").}

\item{dfmax}{An integer which, if not NULL, indicates how many features to be 
extracted (NULL by default).}
}
\value{
If type = "select" the function returns a dataframe with the extracted features, 
else if type = "predict" it returns the vector of predicted labels.
}
\description{
This function uses cross-validated Lasso regression (generalised to multinomial output) 
for both multiclass classification and feature selection.
}
\examples{
# create sample dataset
features <- matrix(rnorm(200), ncol=2)
labels <- c(rep(1, 40), rep(2, 40), rep(3,40), rep(4, 40), rep(5,40))
lasso.model(train.features = features, labels, type = "select", dfmax = 1) #this will select the most important feature in the dataset.
}

