% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.boosting.R
\name{my.boosting}
\alias{my.boosting}
\title{Boosting with adabag}
\usage{
my.boosting(x, y, xtest, mfinal, seed = 1234)
}
\arguments{
\item{x}{A dataframe containing the features for training.}

\item{y}{The labels from the training dataset.}

\item{xtest}{The test dataset for which we want to predict the labels.}

\item{mfinal}{An integer, the number of iterations for which boosting is run or, equivalently, 
the number of trees to use.}

\item{seed}{The initial seed, by default is 1234.}
}
\value{
The predicted labels for the test set.
}
\description{
This function simply uses boosting from 'adabag' package for training the model and testing 
it on the given test set (needed for the rolling windows function).
}

