% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.forest.R
\name{my.forest}
\alias{my.forest}
\title{Random forest with ranger}
\usage{
my.forest(x, y, xtest, ntree = 300, mtry = 4, seed = 1234)
}
\arguments{
\item{x}{A dataframe containing the features for training.}

\item{y}{The labels from the training dataset.}

\item{xtest}{The test dataset for which we want to predict the labels.}

\item{ntree}{An integer, the number of trees to use.}

\item{mtry}{An integer, the depth of each tree.}

\item{seed}{The initial seed, by default is 1234.}
}
\value{
The predicted labels for the test set.
}
\description{
This function uses 'ranger' package to perform a random forest 
(needed for the rolling windows function).
}

