% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.gbm.R
\name{my.gbm}
\alias{my.gbm}
\title{Boosting with gbm}
\usage{
my.gbm(x, y, xtest, seed = 1234)
}
\arguments{
\item{x}{A dataframe containing the features for training.}

\item{y}{The labels from the training dataset.}

\item{xtest}{The test dataset for which we want to predict the labels.}
}
\value{
The predicted labels for the test set.
}
\description{
This function simply uses gbm from 'gbm' package for training the model and testing 
it on the given test set (needed for the rolling windows function).
}

