% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.knn.R
\name{my.knn}
\alias{my.knn}
\title{K-NN with caret}
\usage{
my.knn(x, y, xtest, k)
}
\arguments{
\item{x}{A dataframe containing the features for training.}

\item{y}{The labels from the training dataset.}

\item{xtest}{The test dataset for which we want to predict the labels.}

\item{k}{An integer, the number of nearest neighbors to use.}
}
\value{
The predicted labels for the test set.
}
\description{
This function performs k-NN from 'caret' package (needed for the rolling windows function).
}

