% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.ordinal.R
\name{my.ordinal}
\alias{my.ordinal}
\title{Ordinal regression}
\usage{
my.ordinal(x, y, xtest)
}
\arguments{
\item{x}{A dataframe containing the features for training.}

\item{y}{The labels from the training dataset.}

\item{xtest}{The test dataset for which we want to predict the labels.}
}
\value{
The predicted labels for the test set.
}
\description{
This function performs an ordinal regression of the labels over all the other predictors 
given as inputs (needed for the rolling windows function).
}

