% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.svm.R
\name{my.svm}
\alias{my.svm}
\title{Svm with e1071}
\usage{
my.svm(x, y, xtest, cost, gamma)
}
\arguments{
\item{x}{A dataframe containing the features for training.}

\item{y}{The labels from the training dataset.}

\item{xtest}{The test dataset for which we want to predict the labels.}

\item{cost}{The regularization term for the constraints violation.}

\item{gamma}{Parameter for the formulation of kernels (needed for all except 'linear').}
}
\value{
The predicted labels for the test set.
}
\description{
This function performs SVM using 'e1071' package (needed for the rolling windows function).
}

