% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obtain_date.R
\name{obtain.date}
\alias{obtain.date}
\title{Obtain the date}
\usage{
obtain.date(dataset)
}
\arguments{
\item{dataset}{The Kaggle dataset comprising of training and test data.}
}
\value{
The updated dataset with year, month and is.holiday variables added.
}

