% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictLabels.R
\name{predictLabels}
\alias{predictLabels}
\title{Predict final labels through text mining}
\usage{
predictLabels(time.data)
}
\arguments{
\item{time.data}{The dataframe containing the train and test data}
}
\value{
A dataframe containing the predictions (id and popularity)
}
\description{
This function predicts the final popularity of some articles based on 
the strings contained in the different url's. It obtains the words that appear 
in both train and test data, then it applies some common words
removal (most common words in english, verbs, and adjectives) and finally, removal of those 
words that appear in more than one class. After this, it returns a data frame 
containing all the test data and its predicted popularity (NA for the articles for which 
prediction could not be made, as e.g., no matched words were found in the url's).
}
\examples{
# create sample dataset
time.data <- dataset[order(dataset$timedelta,decreasing = TRUE),]
# obtain the predicted popularity for the test data
predictLabels(time.data)
}

