% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rolling.windows.R
\name{rolling.windows.rf}
\alias{rolling.windows.rf}
\title{Rolling windows with random forest}
\usage{
rolling.windows.rf(dataset, window.size = 10000, step.size = 1000,
  seed = 1234, ntree = 300)
}
\arguments{
\item{dataset}{A dataset comprising of training and test data}

\item{window.size}{An integer which establishes the size of each chunk of the rolling windows.}

\item{step.size}{An integer which defines the distance between the starting observations for two consecutive windows.}

\item{seed}{The seed you want to set for the iterations of the random forest.}

\item{ntree}{The number of trees in each call of random forest.}
}
\value{
A dataframe containing the id's and the predicted labels in the right format for submission.
}
\description{
This function performs random forest with a rolling windows technique, i.e., 
the dataset is split in chunks of fixed size, were each chunk starts from the observation
which follows the starting observation of the previous chunk 
(this is why is called rolling windows).
In each split of the data we predict a portion of the test data labels through a random forest.
The final prediction is given then by a majority vote over all the predictions made.
}

