% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{standardize}
\alias{standardize}
\title{Feature standardization}
\usage{
standardize(x, type = "continuous")
}
\arguments{
\item{x}{A vector containing the feature to be standardized.}

\item{type}{A string with values "continuous" (counts included) or "discrete" (e.g. binary, categorical) depending on the type of the feature.}
}
\value{
The standardized variable.
}
\description{
This function standardizes a feature, either continuous or discrete. 
If it is continuous it applies the normal standardization formula (x - sample mean)/(sample sd), 
else if it is discrete it returns (x-min(x))/(max(x) - min(x)).
}
\examples{
#create a vector
x <- 1:100
x_stand <- standardize(x, type = "continuous") #this will return the standardized feature.
}

