% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Submission.R
\name{submit}
\alias{submit}
\title{Submission generator.}
\usage{
submit(fit, sub_numb)
}
\arguments{
\item{fit}{A vector of length 9644 containing the predicted labels for the test dataset.}

\item{sub_numb}{An integer indicating the number of the generating submission.}
}
\value{
A csv file in the working directory ready for submission named submitX.csv, with X=sub_numb.
}
\description{
Function for creating a csv in the right format for submission.
}

