% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obtain_date.R
\name{obtain.date}
\alias{obtain.date}
\title{This function creates the variables year, month, day and date from the url of each news, 
and also a variable called is.holiday which is a binary that checks if the date is among 
the holiday's days.}
\usage{
obtain.date(dataset)
}
\arguments{
\item{dataset}{The Kaggle dataset comprising of training and test data.}
}
\value{
The updated dataset with year, month and is.holiday variables added.
}
\description{
This function creates the variables year, month, day and date from the url of each news, 
and also a variable called is.holiday which is a binary that checks if the date is among 
the holiday's days.
}

