% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{standardize}
\alias{standardize}
\title{This function standardized a feature, either continuous or discrete.}
\usage{
standardize(x, type = "continuous")
}
\arguments{
\item{x}{A vector containing the feature to be standardized.}

\item{type}{A string with values "continuous" or "discrete" depending on the type of the feature.}
}
\value{
The standardized variable.
}
\description{
This function standardized a feature, either continuous or discrete.
}

